"Fix" clippy checks in CI, resolve a bunch of clippy warnings (#1134)
* Fix a bunch of `clippy` warnings in the `esp-hal` build script

* Resolve most clippy warnings for ESP32 (no features)

* Do the same for ESP32-S2/S3...

* Same for C2/C3...

* CI check for clippy now actually works (and passes!) for Xtensa

* RISC-V chips (except P4) now pass clippy in CI

* Add some safety doc comments which I forgot to write

* The `PwmPeripheral` trait is no longer `unsafe`, and is instead now sealed

+++ esp-hal/src/twai/filter.rs
@@ -255,8 +255,8 @@ impl SingleExtendedFilter {
         let mut acceptance_mask: u32 = 0;
 
         // Pack the id into the full layout.
-        acceptance_code |= (id_code.as_raw() as u32) << 3;
-        acceptance_mask |= (id_mask.as_raw() as u32) << 3;
+        acceptance_code |= id_code.as_raw() << 3;
+        acceptance_mask |= id_mask.as_raw() << 3;
 
         // Pack the rtr bit into the full layout.
         acceptance_code |= (rtr_code as u32) << 2;
